tests/test_pnr_status_checker.py
# tests/test_pnr_status_checker.py  

import pytest  
from pnr_status_checker import PNRStatusChecker  
from playwright.sync_api import sync_playwright  
from unittest.mock import Mock  

@pytest.fixture  
def mock_tesseract():  
    mock = Mock()  
    mock.extract_text_from_image.return_value = "Extracted Text"  
    return mock  

@pytest.fixture  
def mock_playwright():  
    mock = Mock()  
    mock.launch.return_value = Mock()  
    mock.new_page.return_value = Mock()  
    return mock  

def test_pnr_status_checker(mock_tesseract, mock_playwright):  
    pnr_status_checker = PNRStatusChecker("/usr/bin/tesseract")  
    pnr_status_checker.image_processor.extract_text_from_image = mock_tesseract.extract_text_from_image  
    pnr = "8426492816"  
    details = pnr_status_checker.fetch_pnr_status(pnr)  
    assert details is not None  

def test_generate_html_report():  
    details = {  
        'Train Number': '12345',  
        'Train Name': 'Train Name',  
        'Boarding Date (DD-MM-YYYY)': '01-01-2022',  
        'From': 'Source',  
        'To': 'Destination',  
        'Reserved Upto': 'Destination',  
        'Boarding Point': 'Source',  
        'Class': 'Second Class',  
        'S. No.': '1',  
        'Booking Status (Coach No , Berth No., Quota)': 'Booking Status',  
        'Current Status (Coach No , Berth No.)': 'Current Status',  
        'Coach Position': 'Coach Position'  
    }  
    generate_html_report(details)  
    assert True  

def test_locator_iterator():  
    with sync_playwright() as p:  
        browser = p.chromium.launch(headless=True)  
        page = browser.new_page()  
        page.goto("https://www.example.com")  
        locator = page.locator("h1")  
        assert locator.count() > 0  

    browser.close()  

def test_image_processor():  
    image_processor = ImageProcessor("/usr/bin/tesseract")  
    image_path = "path_to_your_image.png"  
    result = image_processor.extract_text_from_image(image_path)  
    assert result is not None  

# Process to test:  
# 1. Run 'pip install pytest' in your terminal.  
# 2. Run 'pip install playwright' in your terminal.  
# 3. Run 'playwright install' in your terminal.  
# 4. Replace 'YOUR_PNR_NUMBER' with your actual PNR number.  
# 5. Run 'pytest tests/test_pnr_status_checker.py' in your terminal.