requirements.txt
greenlet==3.1.1
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
playwright==1.49.1
pyee==12.0.0
pytesseract==0.3.13
typing_extensions==4.12.2
